planner_backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
python-decouple==3.8

